Moved DoubleDataPoint to datapoints package
