+++ common/src/main/kotlin/io/rightpad/daxplorer/data/datapoints/absolute/DoubleDataPoint.kt
@@ -1,4 +1,4 @@
-package io.rightpad.daxplorer.data
+package io.rightpad.daxplorer.data.datapoints.absolute
 
 import io.rightpad.daxplorer.data.datapoints.absolute.TimeSeriesDataPoint
 import java.time.LocalDateTime
