+++ common/src/main/kotlin/io/rightpad/daxplorer/data/features/RelativeStrengthIndex.kt
@@ -1,10 +1,9 @@
 package io.rightpad.daxplorer.data.features
 
-import io.rightpad.daxplorer.data.DoubleDataPoint
+import io.rightpad.daxplorer.data.datapoints.absolute.DoubleDataPoint
 import io.rightpad.daxplorer.data.datapoints.absolute.IndexDataPoint
 import java.time.LocalDateTime
 import java.time.temporal.ChronoUnit
-import java.time.temporal.TemporalUnit
 
 class RelativeStrengthIndex: Feature<DoubleDataPoint> {
     override var featureData: List<DoubleDataPoint>? = null
